// src/app/pages/cookies/cookies-types.ts
type ContentBlock = {
    title: string;
    text: string;
};

export type CookiesContent = {
    pageTitle: string;
    introText: string;
    block: ContentBlock;
    block2: ContentBlock;
    list1: ContentBlock[];
    cookiesLink: string;
    block3: ContentBlock;
    block4: ContentBlock;
    title: string;
    list2: string[];
    buttonText: string;
};